backend/requirements.txt
Flask==2.0.1
python-dotenv==0.19.1
openai==0.27.0
flask-cors==3.0.10
gunicorn==20.1.0